+++ pkg/analyzer/lib/src/test_utilities/find_element.dart
@@ -216,19 +216,22 @@ class FindElement extends _FindElementBase {
   }
 
   CompilationUnitElement part(String targetUri) {
-    CompilationUnitElement? partElement;
+    CompilationUnitElement? result;
 
-    for (var part in libraryElement.parts) {
-      if (part.uri == targetUri) {
-        if (partElement != null) {
-          throw StateError('Not unique: $targetUri');
+    for (final part in libraryElement.parts2) {
+      if (part is PartElementWithPart) {
+        final unitElement = part.includedUnit;
+        if ('${unitElement.source.uri}' == targetUri) {
+          if (result != null) {
+            throw StateError('Not unique: $targetUri');
+          }
+          result = unitElement;
         }
-        partElement = part;
       }
     }
 
-    if (partElement != null) {
-      return partElement;
+    if (result != null) {
+      return result;
     }
     throw StateError('Not found: $targetUri');
   }