+++ pkg/analyzer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.3.0-dev
 * Deprecated `Directive.keyword`, use corresponding `xyzToken` in specific directives.
+* Deprecated `LibraryElement.parts`, use `parts2` instead.
 
 ## 4.2.0
 * Update SDK constraints to `>=2.17.0 <3.0.0`.